feat: fixed bug with custom amount changing to max_uint if toggle was set to max and switched tabs
